requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.50.0
pyee==12.1.1
Pygments==2.19.1
python-dotenv==1.1.0
typing_extensions==4.13.1
Werkzeug==3.1.3
gunicorn==23.0.0